added coordinator state
